+++ src/main/java/zd/zero/waifu/motivator/plugin/listeners/TaskListener.kt
@@ -4,10 +4,10 @@ import com.intellij.openapi.project.ProjectManager
 import com.intellij.task.ProjectTaskListener
 import com.intellij.task.ProjectTaskManager
 import zd.zero.waifu.motivator.plugin.alert.AlertConfiguration
-import zd.zero.waifu.motivator.plugin.assets.VisualMotivationAssetProvider
 import zd.zero.waifu.motivator.plugin.assets.WaifuAssetCategory
-import zd.zero.waifu.motivator.plugin.motivation.VisualMotivationFactory
+import zd.zero.waifu.motivator.plugin.onboarding.UpdateNotification
 import zd.zero.waifu.motivator.plugin.settings.WaifuMotivatorPluginState
+import zd.zero.waifu.motivator.plugin.tools.AssetTools
 
 internal enum class TaskStatus {
     PASS, FAIL, UNKNOWN
@@ -18,28 +18,41 @@ class TaskListener : ProjectTaskListener {
     private var previousTaskStatus = TaskStatus.UNKNOWN
 
     override fun finished(result: ProjectTaskManager.Result) {
+        val project = ProjectManager.getInstance().defaultProject
         when {
             result.hasErrors() -> {
-                VisualMotivationFactory.constructMotivation(
-                    ProjectManager.getInstance().defaultProject,
-                    VisualMotivationAssetProvider.pickAssetFromCategories(
-                        WaifuAssetCategory.DISAPPOINTMENT,
-                        WaifuAssetCategory.SHOCKED
-                    ),
-                    createAlertConfiguration()
-                ).motivate()
+                AssetTools.attemptToShowCategories(
+                    project,
+                    { createAlertConfiguration() },
+                    {
+                        UpdateNotification.sendMessage(
+                            "'Task Failure Motivation' Unavailable Offline",
+                            "Unfortunately I wasn't able to find any waifu saved locally. Please try again " +
+                                "when you are back online!",
+                            project
+                        )
+                    },
+                    WaifuAssetCategory.DISAPPOINTMENT,
+                    WaifuAssetCategory.SHOCKED
+                )
                 previousTaskStatus = TaskStatus.FAIL
             }
             previousTaskStatus == TaskStatus.FAIL -> {
-                VisualMotivationFactory.constructMotivation(
-                    ProjectManager.getInstance().defaultProject,
-                    VisualMotivationAssetProvider.pickAssetFromCategories(
-                        WaifuAssetCategory.CELEBRATION,
-                        WaifuAssetCategory.CELEBRATION,
-                        WaifuAssetCategory.SMUG
-                    ),
-                    createAlertConfiguration()
-                ).motivate()
+                AssetTools.attemptToShowCategories(
+                    project,
+                    { createAlertConfiguration() },
+                    {
+                        UpdateNotification.sendMessage(
+                            "'Task Success Motivation' Unavailable Offline",
+                            "Unfortunately I wasn't able to find any waifu saved locally. Please try again " +
+                                "when you are back online!",
+                            project
+                        )
+                    },
+                    WaifuAssetCategory.CELEBRATION,
+                    WaifuAssetCategory.CELEBRATION,
+                    WaifuAssetCategory.SMUG
+                )
                 previousTaskStatus = TaskStatus.PASS
             }
             else -> {
